Refactor: fix folder names
fix folder user_management, data_sources and use_cases
